requirements.txt
fastapi
uvicorn
pyjwt
pymysql
pydantic
bcrypt
boto3
urllib
cryptography
pillow